Use interaction for SessionDataTask response
When using a delegate it is common to get the response from the
NSURLSessionTask in `URLSession:task:didCompleteWithError:`. However,
the response is never set on a SessionDataTask object so we must
override it with the interaction response.

Unfortunately this means we have to make the `interaction` a property
of SessionDataTask which adds state but the `response` is unavailable
otherwise.