Remove invalid override in extension
Swift 4 complains that properties and methods cannot be overridden in
extensions. I’m not sure if this ever worked before but regardless it’s
now invalid and needs to be removed.

Instead of using the override to hopefully get this behavior when
calling `dictionary` on an `HTTPURLResponse`, we’ll just check to see
if the response is an `HTTPURLResponse` and add the appropriate values
when converting the response into a dictionary in
`Interaction.dictionary`.