Store headers and status code when recording
When finished recording, `Session.persist` calls down to
`Interaction.dictionary`. This getter checks if `response` is a
`HTTPURLResponse`, in which case it records the headers and status code
for the response. However, due to ambiguous namespacing it was checking
if the response was a `DVR.HTTPURLResponse`, not a
`Foundation.HTTPURLResponse`, meaning that recorded HTTP fields would
never be recorded.